smite: move ChannelTypeVariant to bolt::types
The `accept_channel` oracle needs to verify that `channel_type`
uses the smallest possible bitmap and represents a defined
channel type. Keeping `ChannelTypeVariant` in `bolt::types`
gives the oracle a canonical set of channel types to validate
against and avoids duplicating the definition across modules.
