+++ smite-ir/src/operation.rs
@@ -14,7 +14,7 @@ use std::fmt::Write;
 use bitcoin::{opcodes::all as opcodes, script::Builder, script::PushBytes};
 use rand::{Rng, RngExt};
 use serde::{Deserialize, Serialize};
-use smite::bolt::{FeatureBit, Features, ShortChannelId};
+use smite::bolt::{ChannelTypeVariant, ShortChannelId};
 
 use super::VariableType;
 
@@ -420,200 +420,6 @@ impl fmt::Display for ShutdownScriptVariant {
             }
             Self::OpReturn(data) => write!(f, "OpReturn({})", format_hex(data)),
         }
-    }
-}
-
-/// A specific BOLT 2 `channel_type` feature-bit combination.
-///
-/// Each variant corresponds to a channel type accepted by at least one target
-/// implementation:
-///
-/// - `option_static_remotekey` (bit 12)
-/// - `option_anchors` (bits 22 and 12)
-/// - `zero_fee_commitments` (bit 40)
-/// - `option_simple_taproot` (bit 80)
-/// - `option_simple_taproot_staging` (bit 180)
-/// - `option_script_enforced_lease` (bits 2022, 22, 12)
-///
-/// Additionally, the following bits can be added to any channel type:
-/// - `option_scid_alias` (bit 46)
-/// - `option_zeroconf` (bit 50)
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
-pub enum ChannelTypeVariant {
-    /// bit 12
-    StaticRemoteKey,
-    /// bits 12, 46
-    StaticRemoteKeyScidAlias,
-    /// bits 12, 50
-    StaticRemoteKeyZeroConf,
-    /// bits 12, 46, 50
-    StaticRemoteKeyScidAliasZeroConf,
-    /// bits 12, 22
-    Anchors,
-    /// bits 12, 22, 46
-    AnchorsScidAlias,
-    /// bits 12, 22, 50
-    AnchorsZeroConf,
-    /// bits 12, 22, 46, 50
-    AnchorsScidAliasZeroConf,
-    /// bit 40
-    ZeroFeeCommitments,
-    /// bits 40, 46
-    ZeroFeeCommitmentsScidAlias,
-    /// bits 40, 50
-    ZeroFeeCommitmentsZeroConf,
-    /// bits 40, 46, 50
-    ZeroFeeCommitmentsScidAliasZeroConf,
-    /// bit 80
-    SimpleTaproot,
-    /// bits 80, 46
-    SimpleTaprootScidAlias,
-    /// bits 80, 50
-    SimpleTaprootZeroConf,
-    /// bits 80, 46, 50
-    SimpleTaprootScidAliasZeroConf,
-    /// bit 180
-    SimpleTaprootStaging,
-    /// bits 180, 46
-    SimpleTaprootStagingScidAlias,
-    /// bits 180, 50
-    SimpleTaprootStagingZeroConf,
-    /// bits 180, 46, 50
-    SimpleTaprootStagingScidAliasZeroConf,
-    /// bits 12, 22, 2022
-    ScriptEnforcedLease,
-    /// bits 12, 22, 2022, 46
-    ScriptEnforcedLeaseScidAlias,
-    /// bits 12, 22, 2022, 50
-    ScriptEnforcedLeaseZeroConf,
-    /// bits 12, 22, 2022, 46, 50
-    ScriptEnforcedLeaseScidAliasZeroConf,
-}
-
-impl ChannelTypeVariant {
-    /// All variants. Keep in sync with the enum definition.
-    pub const ALL: &[Self] = &[
-        Self::StaticRemoteKey,
-        Self::StaticRemoteKeyScidAlias,
-        Self::StaticRemoteKeyZeroConf,
-        Self::StaticRemoteKeyScidAliasZeroConf,
-        Self::Anchors,
-        Self::AnchorsScidAlias,
-        Self::AnchorsZeroConf,
-        Self::AnchorsScidAliasZeroConf,
-        Self::ZeroFeeCommitments,
-        Self::ZeroFeeCommitmentsScidAlias,
-        Self::ZeroFeeCommitmentsZeroConf,
-        Self::ZeroFeeCommitmentsScidAliasZeroConf,
-        Self::SimpleTaproot,
-        Self::SimpleTaprootScidAlias,
-        Self::SimpleTaprootZeroConf,
-        Self::SimpleTaprootScidAliasZeroConf,
-        Self::SimpleTaprootStaging,
-        Self::SimpleTaprootStagingScidAlias,
-        Self::SimpleTaprootStagingZeroConf,
-        Self::SimpleTaprootStagingScidAliasZeroConf,
-        Self::ScriptEnforcedLease,
-        Self::ScriptEnforcedLeaseScidAlias,
-        Self::ScriptEnforcedLeaseZeroConf,
-        Self::ScriptEnforcedLeaseScidAliasZeroConf,
-    ];
-
-    /// The feature bits (even/required) contained in this channel type.
-    #[must_use]
-    pub fn bits(self) -> &'static [FeatureBit] {
-        use Features as F;
-        match self {
-            Self::StaticRemoteKey => &[F::OPTION_STATIC_REMOTEKEY],
-            Self::StaticRemoteKeyScidAlias => &[F::OPTION_STATIC_REMOTEKEY, F::OPTION_SCID_ALIAS],
-            Self::StaticRemoteKeyZeroConf => &[F::OPTION_STATIC_REMOTEKEY, F::OPTION_ZEROCONF],
-            Self::StaticRemoteKeyScidAliasZeroConf => &[
-                F::OPTION_STATIC_REMOTEKEY,
-                F::OPTION_SCID_ALIAS,
-                F::OPTION_ZEROCONF,
-            ],
-            Self::Anchors => &[F::OPTION_STATIC_REMOTEKEY, F::OPTION_ANCHORS],
-            Self::AnchorsScidAlias => &[
-                F::OPTION_STATIC_REMOTEKEY,
-                F::OPTION_ANCHORS,
-                F::OPTION_SCID_ALIAS,
-            ],
-            Self::AnchorsZeroConf => &[
-                F::OPTION_STATIC_REMOTEKEY,
-                F::OPTION_ANCHORS,
-                F::OPTION_ZEROCONF,
-            ],
-            Self::AnchorsScidAliasZeroConf => &[
-                F::OPTION_STATIC_REMOTEKEY,
-                F::OPTION_ANCHORS,
-                F::OPTION_SCID_ALIAS,
-                F::OPTION_ZEROCONF,
-            ],
-            Self::ZeroFeeCommitments => &[F::ZERO_FEE_COMMITMENTS],
-            Self::ZeroFeeCommitmentsScidAlias => &[F::ZERO_FEE_COMMITMENTS, F::OPTION_SCID_ALIAS],
-            Self::ZeroFeeCommitmentsZeroConf => &[F::ZERO_FEE_COMMITMENTS, F::OPTION_ZEROCONF],
-            Self::ZeroFeeCommitmentsScidAliasZeroConf => &[
-                F::ZERO_FEE_COMMITMENTS,
-                F::OPTION_SCID_ALIAS,
-                F::OPTION_ZEROCONF,
-            ],
-            Self::SimpleTaproot => &[F::OPTION_SIMPLE_TAPROOT],
-            Self::SimpleTaprootScidAlias => &[F::OPTION_SIMPLE_TAPROOT, F::OPTION_SCID_ALIAS],
-            Self::SimpleTaprootZeroConf => &[F::OPTION_SIMPLE_TAPROOT, F::OPTION_ZEROCONF],
-            Self::SimpleTaprootScidAliasZeroConf => &[
-                F::OPTION_SIMPLE_TAPROOT,
-                F::OPTION_SCID_ALIAS,
-                F::OPTION_ZEROCONF,
-            ],
-            Self::SimpleTaprootStaging => &[F::OPTION_SIMPLE_TAPROOT_STAGING],
-            Self::SimpleTaprootStagingScidAlias => {
-                &[F::OPTION_SIMPLE_TAPROOT_STAGING, F::OPTION_SCID_ALIAS]
-            }
-            Self::SimpleTaprootStagingZeroConf => {
-                &[F::OPTION_SIMPLE_TAPROOT_STAGING, F::OPTION_ZEROCONF]
-            }
-            Self::SimpleTaprootStagingScidAliasZeroConf => &[
-                F::OPTION_SIMPLE_TAPROOT_STAGING,
-                F::OPTION_SCID_ALIAS,
-                F::OPTION_ZEROCONF,
-            ],
-            Self::ScriptEnforcedLease => &[
-                F::OPTION_STATIC_REMOTEKEY,
-                F::OPTION_ANCHORS,
-                F::OPTION_SCRIPT_ENFORCED_LEASE,
-            ],
-            Self::ScriptEnforcedLeaseScidAlias => &[
-                F::OPTION_STATIC_REMOTEKEY,
-                F::OPTION_ANCHORS,
-                F::OPTION_SCRIPT_ENFORCED_LEASE,
-                F::OPTION_SCID_ALIAS,
-            ],
-            Self::ScriptEnforcedLeaseZeroConf => &[
-                F::OPTION_STATIC_REMOTEKEY,
-                F::OPTION_ANCHORS,
-                F::OPTION_SCRIPT_ENFORCED_LEASE,
-                F::OPTION_ZEROCONF,
-            ],
-            Self::ScriptEnforcedLeaseScidAliasZeroConf => &[
-                F::OPTION_STATIC_REMOTEKEY,
-                F::OPTION_ANCHORS,
-                F::OPTION_SCRIPT_ENFORCED_LEASE,
-                F::OPTION_SCID_ALIAS,
-                F::OPTION_ZEROCONF,
-            ],
-        }
-    }
-
-    /// Encodes the channel type as a BOLT feature bitmap (big-endian bytes).
-    #[must_use]
-    pub fn encode(self) -> Vec<u8> {
-        Features::from_bits(self.bits()).into_bytes()
-    }
-}
-
-impl fmt::Display for ChannelTypeVariant {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{self:?}")
     }
 }
